BREAKING change: restructure project files ( SWRD-3 )
New project structure takes a huge inspiration from Dimas' and team
project structure. Although this may not be a one-fits-all solution, but
at the very least we can separate between the content and logic side of
the program.
