+++ SwiftRide/Core/Components/SearchBar/BaseSearchBar.swift
@@ -1,0 +1,35 @@
+import SwiftUI
+
+struct SearchBar: View {
+    @Binding var searchText: String
+    @Binding var busStops: [BusStop]
+    
+    var body: some View {
+        VStack (spacing: 10) {
+            HStack {
+                Image(systemName: "magnifyingglass")
+                    .foregroundColor(.gray)
+                
+                TextField("Search Bus Stop", text: $searchText)
+                    .textFieldStyle(PlainTextFieldStyle())
+                
+                if !searchText.isEmpty {
+                    Button {
+                        searchText = ""
+                    } label: {
+                        Image(systemName: "xmark.circle.fill")
+                            .foregroundColor(.gray)
+                    }
+                    
+                }
+            }
+            .padding()
+            .frame(height: 35)
+            .background(Color(.secondarySystemBackground))
+            .clipShape(RoundedRectangle(cornerRadius: 12))
+            .padding(.horizontal)
+            .padding(.top, 25)
+        }
+        .frame( alignment: .top)
+    }
+}
