+++ SwiftRide/Pages/Map/MapViewModel.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct MapViewFoundation { 
+
+}
