+++ SwiftRide/Pages/Sheet/Base/BaseSheetViewModel.swift
@@ -1,0 +1,5 @@
+import SwiftUI
+
+struct BaseSheetViewModel { 
+
+}
