+++ SwiftRide/Pages/Sheet/BusStop/BusStopView.swift
@@ -1,87 +1,5 @@
 import SwiftUI
 
-struct DefaultSheetView: View {
-    @Binding var busStops: [BusStop]
-    @Binding var searchText: String
-    @Binding var selectionDetent: PresentationDetent
-
-    @Binding var selectedSheet: SheetContentType
-
-    @Binding var selectedBusStop: BusStop
-    
-    var body: some View {
-        SearchBar(searchText: $searchText, busStops: $busStops)
-        ScrollView {
-            switch searchText {
-            case "":
-                Text("Nearest Bus Stops...?")
-                Text("Still cooking... 🍳")
-                Text("")
-                
-            default:
-                VStack(spacing: 10) {
-                    ForEach(busStops) { stop in
-                        if stop.name.localizedCaseInsensitiveContains(searchText) || searchText.isEmpty {
-                            HStack {
-                                Circle()
-                                    .frame(width: 30, height: 30)
-                                    .padding(.horizontal, 10)
-                                VStack(alignment: .leading) {
-                                    Text(stop.name)
-                                        .padding(.horizontal, 10)
-                                        .onTapGesture {
-                                            selectedBusStop = stop
-                                            withAnimation(.easeInOut(duration: 0.7)){
-                                                selectedSheet = .busStopDetailView
-                                                selectionDetent = .medium
-                                            }
-                                        }
-                                }
-                                Spacer()
-                            }
-                        }
-                    }
-                    .padding()
-                }
-            }
-        }
-    }
-}
-
-struct SearchBar: View {
-    @Binding var searchText: String
-    @Binding var busStops: [BusStop]
-    
-    var body: some View {
-        VStack (spacing: 10) {
-            HStack {
-                Image(systemName: "magnifyingglass")
-                    .foregroundColor(.gray)
-                
-                TextField("Search Bus Stop", text: $searchText)
-                    .textFieldStyle(PlainTextFieldStyle())
-                
-                if !searchText.isEmpty {
-                    Button {
-                        searchText = ""
-                    } label: {
-                        Image(systemName: "xmark.circle.fill")
-                            .foregroundColor(.gray)
-                    }
-                    
-                }
-            }
-            .padding()
-            .frame(height: 35)
-            .background(Color(.secondarySystemBackground))
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-            .padding(.horizontal)
-            .padding(.top, 25)
-        }
-        .frame( alignment: .top)
-    }
-}
-
 struct BusStopDetailView: View {
     @Binding var currentBusStop: BusStop
     @Binding var selectedSheet: SheetContentType
