--- SwiftRide/Schedule/Schedule_Breeze.json
@@ -1,173 +1,0 @@
-[
-  { "bus_stop_name": "The Breeze",
-    "bus_number": "7",
-    "time_of_arrival": "10.00",
-    "session": "2"
-  },
-  { "bus_stop_name": "CBD Timur 1",
-    "bus_number": "7",
-    "time_of_arrival": "10.02",
-    "session": "2"
-  },
-  { "bus_stop_name": "Lobby AEON Mall",
-    "bus_number": "7",
-    "time_of_arrival": "10.09",
-    "session": "2"
-  },
-  { "bus_stop_name": "AEON Mall 2",
-    "bus_number": "7",
-    "time_of_arrival": "10.11",
-    "session": "2"
-  },
-  { "bus_stop_name": "CBD Utara 3",
-    "bus_number": "7",
-    "time_of_arrival": "10.12",
-    "session": "2"
-  },
-  { "bus_stop_name": "ICE 1",
-    "bus_number": "7",
-    "time_of_arrival": "10.15",
-    "session": "2"
-  },
-  { "bus_stop_name": "ICE 2,
-    "bus_number": "7",
-    "time_of_arrival": "10.20",
-    "session": "2"
-  },
-  { "bus_stop_name": "ICE 6",
-    "bus_number": "7",
-    "time_of_arrival": "10.22",
-    "session": "2"
-  },
-  { "bus_stop_name": "ICE 5",
-    "bus_number": "7",
-    "time_of_arrival": "10.23",
-    "session": "2"
-  },
-  { "bus_stop_name": "CBD Barat 1",
-    "bus_number": "7",
-    "time_of_arrival": "10.26",
-    "session": "2"
-  },
-  { "bus_stop_name": "CBD Barat 2",
-    "bus_number": "7",
-    "time_of_arrival": "10.26",
-    "session": "2"
-  },
-  { "bus_stop_name": "Lobby AEON Mall",
-    "bus_number": "7",
-    "time_of_arrival": "10.33",
-    "session": "2"
-  },
-  { "bus_stop_name": "AEON Mall 2",
-    "bus_number": "7",
-    "time_of_arrival": "10.34",
-    "session": "2"
-  },
-  { "bus_stop_name": "CBD Timur 2",
-    "bus_number": "7",
-    "time_of_arrival": "10.37",
-    "session": "2"
-  },
-  { "bus_stop_name": "Nava Park 1",
-    "bus_number": "7",
-    "time_of_arrival": "10.39",
-    "session": "2"
-  },
-  { "bus_stop_name": "Green Cove",
-    "bus_number": "7",
-    "time_of_arrival": "10.42",
-    "session": "2"
-  },
-  { "bus_stop_name": "The Breeze",
-    "bus_number": "7",
-    "time_of_arrival": "10.45",
-    "session": "2"
-  },
--- spacer --
-  { "bus_stop_name": "The Breeze",
-    "bus_number": "7",
-    "time_of_arrival": "17.00",
-    "session": "9"
-  },
-  { "bus_stop_name": "CBD Timur 1",
-    "bus_number": "7",
-    "time_of_arrival": "17.02",
-    "session": "9"
-  },
-  { "bus_stop_name": "Lobby AEON Mall",
-    "bus_number": "7",
-    "time_of_arrival": "17.09",
-    "session": "9"
-  },
-  { "bus_stop_name": "AEON Mall 2",
-    "bus_number": "7",
-    "time_of_arrival": "17.11",
-    "session": "9"
-  },
-  { "bus_stop_name": "CBD Utara 3",
-    "bus_number": "7",
-    "time_of_arrival": "17.12",
-    "session": "9"
-  },
-  { "bus_stop_name": "ICE 1",
-    "bus_number": "7",
-    "time_of_arrival": "17.15",
-    "session": "9"
-  },
-  { "bus_stop_name": "ICE 2,
-    "bus_number": "7",
-    "time_of_arrival": "17.20",
-    "session": "9"
-  },
-  { "bus_stop_name": "ICE 6",
-    "bus_number": "7",
-    "time_of_arrival": "17.22",
-    "session": "9"
-  },
-  { "bus_stop_name": "ICE 5",
-    "bus_number": "7",
-    "time_of_arrival": "17.23",
-    "session": "9"
-  },
-  { "bus_stop_name": "CBD Barat 1",
-    "bus_number": "7",
-    "time_of_arrival": "17.26",
-    "session": "9"
-  },
-  { "bus_stop_name": "CBD Barat 2",
-    "bus_number": "7",
-    "time_of_arrival": "17.26",
-    "session": "9"
-  },
-  { "bus_stop_name": "Lobby AEON Mall",
-    "bus_number": "7",
-    "time_of_arrival": "17.33",
-    "session": "9"
-  },
-  { "bus_stop_name": "AEON Mall 2",
-    "bus_number": "7",
-    "time_of_arrival": "17.34",
-    "session": "9"
-  },
-  { "bus_stop_name": "CBD Timur 2",
-    "bus_number": "7",
-    "time_of_arrival": "17.37",
-    "session": "9"
-  },
-  { "bus_stop_name": "Nava Park 1",
-    "bus_number": "7",
-    "time_of_arrival": "17.39",
-    "session": "9"
-  },
-  { "bus_stop_name": "Green Cove",
-    "bus_number": "7",
-    "time_of_arrival": "17.42",
-    "session": "9"
-  },
-  { "bus_stop_name": "The Breeze",
-    "bus_number": "7",
-    "time_of_arrival": "17.45",
-    "session": "9"
-  },
-]
