--- SwiftRide/Schedule/Schedule_Intermoda_Sektor1.3.json
@@ -1,263 +1,0 @@
-[
-  { "bus_stop_name": "Intermoda",
-    "bus_number": "1",  
-    "time_of_arrival": "06.00",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Cosmo",
-    "bus_number": "1",  
-    "time_of_arrival": "06.04",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Verdant View",
-    "bus_number": "1",  
-    "time_of_arrival": "06.05",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Eternity",
-    "bus_number": "1",  
-    "time_of_arrival": "06.06",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Simplicity 2",
-    "bus_number": "1",  
-    "time_of_arrival": "06.07",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Edutown 1",
-    "bus_number": "1",  
-    "time_of_arrival": "06.09",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Edutown 2",
-    "bus_number": "1",  
-    "time_of_arrival": "06.09",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 1",
-    "bus_number": "1",  
-    "time_of_arrival": "06.10",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 2",
-    "bus_number": "1",  
-    "time_of_arrival": "06.11",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 6",
-    "bus_number": "1",  
-    "time_of_arrival": "06.13",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 5",
-    "bus_number": "1",  
-    "time_of_arrival": "06.14",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "GOP 1",
-    "bus_number": "1",  
-    "time_of_arrival": "06.18",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "SML Plaza",
-    "bus_number": "1",  
-    "time_of_arrival": "06.19",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "The Breeze",
-    "bus_number": "1",  
-    "time_of_arrival": "06.21",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "CBD Timur 1",
-    "bus_number": "1",  
-    "time_of_arrival": "06.23",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "CBD Timur 2",
-    "bus_number": "1",  
-    "time_of_arrival": "06.24",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Nava Park 1",
-    "bus_number": "1",  
-    "time_of_arrival": "06.26",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "SWA 2",
-    "bus_number": "1",  
-    "time_of_arrival": "06.28",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Giant",
-    "bus_number": "1",  
-    "time_of_arrival": "06.30",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Eka Hospital 1",
-    "bus_number": "1",  
-    "time_of_arrival": "06.33",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Puspita Loka",
-    "bus_number": "1",  
-    "time_of_arrival": "06.36",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Polsek Serpong",
-    "bus_number": "1",  
-    "time_of_arrival": "06.38",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Ruko Madrid",
-    "bus_number": "1",  
-    "time_of_arrival": "06.40",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Pasar Modern Timur",
-    "bus_number": "1",  
-    "time_of_arrival": "06.41",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Griya Loka 1",
-    "bus_number": "1",  
-    "time_of_arrival": "06.43",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Halte Sektor 1.3",
-    "bus_number": "1",  
-    "time_of_arrival": "06.45",
-    "session": "1" 
-  }, 
--- spacer -- 
-  { "bus_stop_name": "Intermoda",
-    "bus_number": "2",  
-    "time_of_arrival": "06.30",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Cosmo",
-    "bus_number": "2",  
-    "time_of_arrival": "06.34",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Verdant View",
-    "bus_number": "2",  
-    "time_of_arrival": "06.35",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Eternity",
-    "bus_number": "2",  
-    "time_of_arrival": "06.36",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Simplicity 2",
-    "bus_number": "2",  
-    "time_of_arrival": "06.37",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Edutown 1",
-    "bus_number": "2",  
-    "time_of_arrival": "06.39",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Edutown 2",
-    "bus_number": "2",  
-    "time_of_arrival": "06.39",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 1",
-    "bus_number": "2",  
-    "time_of_arrival": "06.40",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 2",
-    "bus_number": "2",  
-    "time_of_arrival": "06.41",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 6",
-    "bus_number": "2",  
-    "time_of_arrival": "06.43",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "ICE 5",
-    "bus_number": "2",  
-    "time_of_arrival": "06.44",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "GOP 1",
-    "bus_number": "2",  
-    "time_of_arrival": "06.50",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "SML Plaza",
-    "bus_number": "2",  
-    "time_of_arrival": "06.51",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "The Breeze",
-    "bus_number": "2",  
-    "time_of_arrival": "06.53",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "CBD Timur 1",
-    "bus_number": "2",  
-    "time_of_arrival": "06.55",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "CBD Timur 2",
-    "bus_number": "2",  
-    "time_of_arrival": "06.56",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Nava Park 1",
-    "bus_number": "2",  
-    "time_of_arrival": "06.58",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "SWA 2",
-    "bus_number": "2",  
-    "time_of_arrival": "07.00",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Giant",
-    "bus_number": "2",  
-    "time_of_arrival": "07.04",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Eka Hospital 1",
-    "bus_number": "2",  
-    "time_of_arrival": "07.07",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Puspita Loka",
-    "bus_number": "2",  
-    "time_of_arrival": "07.10",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Polsek Serpong",
-    "bus_number": "2",  
-    "time_of_arrival": "07.14",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Ruko Madrid",
-    "bus_number": "2",  
-    "time_of_arrival": "07.17",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Pasar Modern Timur",
-    "bus_number": "2",  
-    "time_of_arrival": "07.19",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Griya Loka 1",
-    "bus_number": "2",  
-    "time_of_arrival": "07.21",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Halte Sektor 1.3",
-    "bus_number": "2",  
-    "time_of_arrival": "07.23",
-    "session": "1" 
-  }, 
-]
