--- SwiftRide/Schedule/Schedule_Vanya.json
@@ -1,213 +1,0 @@
-[
-  { "bus_stop_name": "Terminal Intermoda",
-    "bus_number": "9",
-    "time_of_arrival": "06.00",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Simplicity 2",
-    "bus_number": "9",
-    "time_of_arrival": "06.03",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "Edutown 1",
-    "bus_number": "9",
-    "time_of_arrival": "06.05",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Edutown 2",
-    "bus_number": "9",
-    "time_of_arrival": "06.05",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "ICE 1",
-    "bus_number": "9",
-    "time_of_arrival": "06.06",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "ICE 2",
-    "bus_number": "9",
-    "time_of_arrival": "06.07",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "ICE 6",
-    "bus_number": "1",
-    "time_of_arrival": "06.09",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Prestigia",
-    "bus_number": "1",
-    "time_of_arrival": "06.10",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "The Mozia 1",
-    "bus_number": "9",
-    "time_of_arrival": "06.11",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Piazza Mozia",
-    "bus_number": "9",
-    "time_of_arrival": "06.11",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "Tabebuya",
-    "bus_number": "9",
-    "time_of_arrival": "06.15",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "Vanya Park",
-    "bus_number": "9",
-    "time_of_arrival": "06.17",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "The Mozia 2",
-    "bus_number": "9",
-    "time_of_arrival": "06.21",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "Illustria",
-    "bus_number": "9",
-    "time_of_arrival": "06.22",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "ICE 2",
-    "bus_number": "9",
-    "time_of_arrival": "06.24",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "ICE 6",
-    "bus_number": "9",
-    "time_of_arrival": "06.26",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "ICE 5",
-    "bus_number": "9",
-    "time_of_arrival": "06.27",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "CBD Barat 1",
-    "bus_number": "9",
-    "time_of_arrival": "06.31",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "CBD Barat 2",
-    "bus_number": "9",
-    "time_of_arrival": "06.31",
-    "session": "1"
-  }, 
-  { "bus_stop_name": "Simplicity 1",
-    "bus_number": "9",
-    "time_of_arrival": "06.33",
-    "session": "1" 
-  }, 
-  { "bus_stop_name": "Terminal Intermoda",
-    "bus_number": "9",
-    "time_of_arrival": "06.35",
-    "session": "1"
-  },
--- spacer -- 
- { "bus_stop_name": "Terminal Intermoda",
-   "bus_number": "9",
-   "time_of_arrival": "10.05",
-   "session": "6"
- },
- { "bus_stop_name": "Simplicity 2",
-   "bus_number": "9",
-   "time_of_arrival": "10.08",
-   "session": "6"
- },
- { "bus_stop_name": "Edutown 1",
-   "bus_number": "9",
-   "time_of_arrival": "10.10",
-   "session": "6"
- },
- { "bus_stop_name": "Edutown 2",
-   "bus_number": "9",
-   "time_of_arrival": "10.10",
-   "session": "6"
- },
- { "bus_stop_name": "ICE 1",
-   "bus_number": "9",
-   "time_of_arrival": "10.11",
-   "session": "6"
- },
- { "bus_stop_name": "ICE 2",
-   "bus_number": "9",
-   "time_of_arrival": "10.12",
-   "session": "6"
- },
- { "bus_stop_name": "ICE 6",
-   "bus_number": "1",
-   "time_of_arrival": "10.14",
-   "session": "6"
- },
- { "bus_stop_name": "Prestigia",
-   "bus_number": "1",
-   "time_of_arrival": "10.15",
-   "session": "6"
- },
- { "bus_stop_name": "The Mozia 1",
-   "bus_number": "9",
-   "time_of_arrival": "10.16",
-   "session": "6"
- },
- { "bus_stop_name": "Piazza Mozia",
-   "bus_number": "9",
-   "time_of_arrival": "10.16",
-   "session": "6"
- },
- { "bus_stop_name": "Tabebuya",
-   "bus_number": "9",
-   "time_of_arrival": "10.20",
-   "session": "6"
- },
- { "bus_stop_name": "Vanya Park",
-   "bus_number": "9",
-   "time_of_arrival": "10.22",
-   "session": "6"
- },
- { "bus_stop_name": "The Mozia 2",
-   "bus_number": "9",
-   "time_of_arrival": "10.26",
-   "session": "6"
- },
- { "bus_stop_name": "Illustria",
-   "bus_number": "9",
-  "time_of_arrival": "10.27",
-  "session": "6"
- },
- { "bus_stop_name": "ICE 2",
-   "bus_number": "9",
-   "time_of_arrival": "10.29",
-   "session": "6"
- },
- { "bus_stop_name": "ICE 6",
-   "bus_number": "9",
-   "time_of_arrival": "10.31",
-   "session": "6"
- },
- { "bus_stop_name": "ICE 5",
-   "bus_number": "9",
-   "time_of_arrival": "10.32",
-   "session": "6"
- },
- { "bus_stop_name": "CBD Barat 1",
-   "bus_number": "9",
-   "time_of_arrival": "10.36",
-   "session": "6"
- },
- { "bus_stop_name": "CBD Barat 2",
-   "bus_number": "9",
-   "time_of_arrival": "10.36",
-   "session": "6"
- },
- { "bus_stop_name": "Simplicity 1",
-   "bus_number": "9",
-   "time_of_arrival": "10.38",
-   "session": "6"
- },
- { "bus_stop_name": "Terminal Intermoda",
-   "bus_number": "9",
-   "time_of_arrival": "10.40",
-   "session": "6"
- },
-]
